+++ backend/middleware/auth.js
@@ -1,69 +1,29 @@
 const jwt = require("jsonwebtoken");
+const express = require("express");
+const router = express.Router();
 
-function authenticateToken(req, res, next) {
-  const authHeader = req.headers["authorization"];
-  const token = authHeader && authHeader.split(" ")[1];
-  console.log(token);
-  if (!token) {
-    return res.sendStatus(401);
-  }
-
-  jwt.verify(token, process.env.JWT_SECRET, (err, user) => {
-    if (err) {
-      console.log(err);
-      return res.sendStatus(403);
-    }
-
-    req.user = user;
-    next();
-  });
-}
-
-module.exports = {
-  authenticateToken,
-};
-
-function authenticateToken(req, res, next) {
+const authenticateToken = (req, res, next) => {
   const authHeader = req.headers["authorization"];
   const token = authHeader && authHeader.split(" ")[1];
   if (token == null) return res.sendStatus(401);
 
-  jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
+  jwt.verify(token, process.env.JWT_SECRET, (err, user) => {
     if (err) return res.sendStatus(403);
     req.user = user;
     next();
   });
-}
+};
 
-function authenticateToken(req, res, next) {
-  const token = req.headers.authorization.split(" ")[1]; // Extract token from headers
-
-  if (!token) {
-    return res.status(401).json({ message: "Unauthorized" });
-  }
-
-  try {
-    const decoded = jwt.verify(token, process.env.JWT_SECRET); // Verify and decode token
-    req.user = decoded;
-    next();
-  } catch (err) {
-    return res.status(403).json({ message: "Forbidden" });
-  }
-}
-
-function authenticateToken(req, res, next) {
+router.use((req, res, next) => {
   const authHeader = req.headers["authorization"];
   const token = authHeader && authHeader.split(" ")[1];
-  if (!token) {
-    return res.status(401).json({ message: "Unauthorized" });
-  }
+  if (token == null) return res.sendStatus(401);
 
   jwt.verify(token, process.env.JWT_SECRET, (err, user) => {
-    if (err) {
-      return res.status(403).json({ message: "Forbidden" });
-    }
-
+    if (err) return res.sendStatus(403);
     req.user = user;
     next();
   });
-}
+});
+
+module.exports = authenticateToken;
