+++ backend/src/contestant-routes.js
@@ -6,7 +6,7 @@ const Contestant = require("../models/contestant");
 const express = require("express");
 const router = express.Router();
 
-const { authenticateToken } = require("../middleware/auth");
+const authenticateToken = require("../middleware/auth");
 
 // Contestants Test Objects in MongoDB
 
@@ -29,15 +29,13 @@ router.get("/", authenticateToken, async (req, res) => {
 router.get("/:contestantId", async (req, res) => {
   // --- YOUR CODE GOES UNDER THIS LINE ---
   try {
-      const contestant1 = await Contestant.findById(req.params.contestantId);
-      console.log(contestant1);
-      res.json(contestant1);
+    const contestant1 = await Contestant.findById(req.params.contestantId);
+    console.log(contestant1);
+    res.json(contestant1);
   } catch (err) {
-      console.log(err);
-      res.json({ message: err });
-
+    console.log(err);
+    res.json({ message: err });
   }
-  
 });
 
 // Create a new Contestant object
@@ -59,40 +57,40 @@ router.post("/", async (req, res) => {
 
 // Update a specific Contestant object
 router.patch("/:contestantId", async (req, res) => {
-    try {
-        const contestant1 = await Contestant.findById(req.params.contestantId);
+  try {
+    const contestant1 = await Contestant.findById(req.params.contestantId);
 
-        if (req.body.name) {
-            contestant1.name = req.body.name;
-        }
-
-        if (req.body.active === true) {
-          contestant1.active = req.body.active;
-        }
-
-        else if (req.body.active === false ) {
-          contestant1.active = req.body.active; 
-        }
-
-        if (req.body.school) {
-          contestant1.school = req.body.school;
-        }
-        const savedcontestant1 = await contestant1.save();
-        res.json(contestant1);
-    } catch (err) {
-        res.json({ message: err });
+    if (req.body.name) {
+      contestant1.name = req.body.name;
     }
+
+    if (req.body.active === true) {
+      contestant1.active = req.body.active;
+    } else if (req.body.active === false) {
+      contestant1.active = req.body.active;
+    }
+
+    if (req.body.school) {
+      contestant1.school = req.body.school;
+    }
+    const savedcontestant1 = await contestant1.save();
+    res.json(contestant1);
+  } catch (err) {
+    res.json({ message: err });
+  }
 });
 
 // Delete a specific Contestant object
 router.delete("/:contestantId", async (req, res) => {
   // --- YOUR CODE GOES UNDER THIS LINE ---
   try {
-    const removedContestant = await Contestant.deleteOne({ _id: req.params.contestantId });
+    const removedContestant = await Contestant.deleteOne({
+      _id: req.params.contestantId,
+    });
     res.json(removedContestant);
   } catch (err) {
-      console.log(err);
-      res.json({ message: err });
+    console.log(err);
+    res.json({ message: err });
   }
 });
